+++ bingads/v13/bulk/entities/bulk_offline_conversion.py
@@ -24,7 +24,61 @@ class BulkOfflineConversion(_SingleRecordBulkEntity):
     def __init__(self, offline_conversion=None):
         super(BulkOfflineConversion, self).__init__()
         self._offline_conversion = offline_conversion
-
+        self._adjustment_value = None
+        self._adjustment_time = None
+        self._adjustment_type = None
+        self._adjustment_currency_code = None
+        self._external_attribution_model = None
+        self._external_attribution_credit = None
+        
+    @property
+    def adjustment_value(self):
+        return self._adjustment_value;
+    
+    @adjustment_value.setter
+    def adjustment_value(self, value):
+        self._adjustment_value = value
+        
+    @property
+    def adjustment_time(self):
+        return self._adjustment_time;
+    
+    @adjustment_time.setter
+    def adjustment_time(self, value):
+        self._adjustment_time = value
+ 
+    @property
+    def adjustment_type(self):
+        return self._adjustment_type;
+    
+    @adjustment_type.setter
+    def adjustment_type(self, value):
+        self._adjustment_type = value
+        
+    @property
+    def adjustment_currency_code(self):
+        return self._adjustment_currency_code;
+    
+    @adjustment_currency_code.setter
+    def adjustment_currency_code(self, value):
+        self._adjustment_currency_code = value
+      
+    @property
+    def external_attribution_model(self):
+        return self._external_attribution_model;
+    
+    @external_attribution_model.setter
+    def external_attribution_model(self, value):
+        self._external_attribution_model = value
+        
+    @property
+    def external_attribution_credit(self):
+        return self._external_attribution_credit;
+    
+    @external_attribution_credit.setter
+    def external_attribution_credit(self, value):
+        self._external_attribution_credit = value                     
+        
     @property
     def offline_conversion(self):
         """ The offline conversion Data Object of the Campaign Management Service.
@@ -80,6 +134,65 @@ class BulkOfflineConversion(_SingleRecordBulkEntity):
             csv_to_field=lambda c, v: setattr(
                 c.offline_conversion,
                 'ConversionTime',
+                parse_datetime(v) if v else None
+            )
+        ),
+        _SimpleBulkMapping(
+            header=_StringTable.AdjustmentValue,
+            field_to_csv=lambda c: c.adjustment_value,
+            csv_to_field=lambda c, v: setattr(
+                c,
+                'adjustment_value',
+                float(v) if v else None
+            )
+        ),
+        
+        _SimpleBulkMapping(
+            header=_StringTable.AdjustmentType,
+            field_to_csv=lambda c: c.adjustment_type,
+            csv_to_field=lambda c, v: setattr(
+                c,
+                'adjustment_type',
+                v
+            )
+        ),
+        
+        _SimpleBulkMapping(
+            header=_StringTable.AdjustmentCurrencyCode,
+            field_to_csv=lambda c: c.adjustment_currency_code,
+            csv_to_field=lambda c, v: setattr(
+                c,
+                'adjustment_currency_code',
+                v
+            )
+        ),
+        
+        _SimpleBulkMapping(
+            header=_StringTable.ExternalAttributionModel,
+            field_to_csv=lambda c: c.external_attribution_model,
+            csv_to_field=lambda c, v: setattr(
+                c,
+                'external_attribution_model',
+                v
+            )
+        ),
+        _SimpleBulkMapping(
+            header=_StringTable.ExternalAttributionCredit,
+            field_to_csv=lambda c: c.external_attribution_credit,
+            csv_to_field=lambda c, v: setattr(
+                c,
+                'external_attribution_credit',
+                float(v) if v else None
+            )
+        ),
+        
+        
+        _SimpleBulkMapping(
+            header=_StringTable.AdjustmentTime,
+            field_to_csv=lambda c: bulk_datetime_str(c.adjustment_time),
+            csv_to_field=lambda c, v: setattr(
+                c,
+                'adjustment_time',
                 parse_datetime(v) if v else None
             )
         ),